refactor: Improve test infrastructure and configuration
- Migrate ESLint config to TypeScript for better type safety
- Simplify Turbo cache configuration with optimized inputs
- Enhance test utilities with better provider registry injection
- Standardize test helper patterns across E2E and unit tests
- Improve type safety in test provider factories
