+++ client/eslint.config.ts
